requirements.txt

matplotlib
streamlit
pandas
pyspellchecker
torch
transformers
watchdog